Rename `NoOpCorruptionHandler` to be `ReThrowCorruptionHandler`.
This rename is done to align the naming of the 2 out-of-the-box corruption handlers we have: "No-Op" gives off the impression that the exception would be swallowed somehow, but the `NoOpCorruptionHandler` actually rethrows the exception. This change aligns the naming to make sense with what the corruption handler is doing, similar to how the `ReplaceFileCorruptionHandler` actually replaces the file.

Bug: b/346197747
Test: N/A
Change-Id: Icca60d0398e06db106e92d53b443661a9088c732