Include transitive source sets (like before) for kotlin-project-structure-metadata.json
After converting to a default source set hiearrcy, the kotlin-project-structure-metadata.json stopped including the transitive
source sets. This fixes change fixes that.

BUG: 428757597
TESTED: Generated kotlin-project-structure-metadata.json for a few projects and it was the same as before my change in applying the new default source set,
I also generated docs for the ui-test-junit4 lib and i no longer see the errors that existed in ab/14561861

Change-Id: Id14ddc7130f13327b1efbdf3509226978265659a