Add default to HasDefaultViewModelProviderFactory
The `defaultViewModelProviderFactory` property in the
`HasDefaultViewModelProviderFactory` interface now has a default
implementation returning `DefaultViewModelProviderFactory`.

This allows classes implementing this interface to omit the property
override if the standard internal factory is sufficient.

Note that `DefaultViewModelProviderFactory` is internal, so external
classes could not previously reference it. This change exposes the
instance via the interface default.

RelNote: "Add a default implementation for
`HasDefaultViewModelProviderFactory.defaultViewModelProviderFactory`.
Implementations of this interface no longer need to override this
property explicitly."

Bug: N/A
Test: existing passes
Change-Id: Ia709580f9344ae5e77eabcc840a384c4bd516e1c