Add RotationProvider in camera-core
The RotationProvider will be lazily created when CameraX#getRotationProvider is first called. The underlying orientation listener is enabled only when there are listeners registered to the RotationProvider. Otherwise, it will be disabled.

To avoid rapid rotation changes, a hysteresis mechanism is applied to the RotationProvider in camera-core. When the orientation is close to a boundary, the rotation will not change until it crosses a threshold.

Bug: 467187356
Test: RotationProviderTest, CameraXTest

Change-Id: I9f339f4cf4d89c66cbf78216296e51bf89d5e470