Add capability to opt-out of premium network slicing
This change introduces a new capability,
`CAPABILITY_OPT_OUT_OF_PREMIUM_NETWORK`, in `CallsManager`. This allows
VoIP applications to signal that they want to opt-out of the system's
default behavior of requesting premium network slices for voice and
video calls.

By default, the system may automatically request a premium network slice
to improve the quality of service for calls.

This capability should be passed in `registerAppWithTelecom`.

Bug: 447631226
Bug: 430509489
Test: manual and atest

Change-Id: I2dfdb6499748a48fb728cd92e04158445778b5a8