Fix bug in RemotePathBase and add RemoteIcon samples
This change fixes a bug in `RemotePathBase.rQuadTo` where the x-coordinate was used for a y-coordinate calculation. It also refactors `rMoveTo` to use a dedicated `addMove` method.

Additionally, new samples and previews are added for `RemoteIcon` to demonstrate its usage, including loading an icon from a vector resource. Tests for `RemotePathParser` are also included, using path data from a real vector drawable to ensure correctness.

Bug: 467979849
Test: New unit tests added.
Change-Id: If11b72f58ba2bc61f42ed95943b758455242a9f1