Fix cursor input type clearing in InputDispatcher
Adds a fix for an additional case of maintaining invalid state in
InputDispatcher.

Pressing a mouse/trackpad button exits hover on Android, which was
clearing the cursor input type erroneously.

If in that state, a touch event came to cancel the in-progress
mouse/trackpad gesture, an exception would be thrown.

Tests: New tests

Change-Id: Iabef34b3f7129577fbfc0bc0d07cd4512d113c94