Apply last recommended transform to initial keyEntity.
When a keyEntity is set for the first time, SceneCore invokes the SpatialModeChangedListener with latest available recommended pose and scale values provided by the system.

This ensures the entity appears with the correct transform upon initialization, rather than waiting for a subsequent spatial mode change event. The last recommended values are now cached for this purpose.

Bug: b/468393004
RelNote: Upon setting valid Entity as keyEntity, when it wasn't set earlier, triggers a spatialModeChangedCallback with last seen values.
Test: Added unit tests and tested manually.
Change-Id: I5793aaf42859cd6197ee2dc795c5937a40bfef64