Refactor HostAction to use RemoteString for its name
The `name` property of the `HostAction` class is changed from a `String` to a `RemoteString`. This allows the action's name to be dynamic and remotely updatable.

Constructors and the `toRemoteAction` method are updated to handle the new `RemoteString` type, including a check for constant values to optimize the generated remote action.

The import statements for `MutableRemoteInt` and `MutableRemoteString` have been replaced with `RemoteInt` and `RemoteString` respectively.

Test: Added HostActionTest
Bug: 464464202
Change-Id: I08f852300da59cf9c148807773dd4fb0271a7a9f