Invalidate root layer when all contents are removed
Fixes: 299503084

When the last contents of Compose were being removed, the layer
that contains it wasn't being invalidated. This adds the ability
to invalidate that root layer without adding a superfluous layer
to the root node.

Relnote: "Fixes bug where removing all content from the ComposeView
doesn't remove the drawn content."

Test: new test, manual testing
Change-Id: I8b3820ded1d89a3b4968384e6965ca47075cedff