SurfaceProcessor: Ignore frames arriving after surface release
On some devices, a few frames can still be sent for processing even
after the camera session has been closed and the output Surface has been
released. This can lead to `onFrameAvailable` being called on a released
`SurfaceTexture`, causing errors.

This CL addresses this race condition by:
1.  Updating `ToneMappingSurfaceProcessor` to track released
    `SurfaceTexture` instances and ignore any subsequent frames.
2.  Adding documentation and code sample to the `SurfaceProcessor`
    interface to guide developers in handling this scenario correctly.

This fix allows for the re-enabling of the `EffectsFragmentDeviceTest`,
which was previously disabled due to this issue on certain devices.

Bug: 465855844
Test: EffectsFragmentDeviceTest
Change-Id: I964e8403cfdb6019f09316f515f966126e27f2d6