Update room-compiler to generate web compatible code
Web compatible code is common code that does not rely on non-web APIs, specifically this change updates Room generated code to use the added extension functions in androidx.sqlite. It also updates the generation of various functions to be suspending functions to adjust to the room-runtime changes to support web.

Bug: 336758416
Bug: 438041176
Test: Existing
Change-Id: Ifaba1541b138b25eb5fc5b6d462d7c57399a64b4