Update Surface maps when image sources are present
When all streams are configured with ImageSourceConfigs, we will not get
SurfaceGraph.set() since no CameraGraph.setSurface() is invoked.

Make CameraGraph trigger an initial update call when image sources are
present. This may also help if the non-ImageSourceConfig streams are
configured with deferred outputs. Hence we trigger this call whenever
there are ImageSourceConfig streams, as opposed to just do this when all
of them are ImageSourceConfig streams.

Bug: 470162657
Test: Configure all streams in CameraPipe sample app with ImageSource.
Change-Id: I33aae49d305d84465f06c94386519813aa314f96