Fix the issue that shadows of levitated panes are not clipped
Due to b/375496210, alpha setting doesn't work with shadows and causes
shadows cannot be faded in/out. To workaround this issue, in our
AnimatedPane implementation we have to apply shadow on
AnimatedVisibility instead of the actual pane container. However, if any
graphic layer transformation is applied to the pane, for example
clipping, this will result in inconsistencies between pane rendering and
its shadow.

This CL fixes/workarounds this by extracting GraphicsLayerElements from
the provided modifier and apply them to AnimatedVisibility instead, to
ensure the consistency between panes and their shadows. Note that
GraphicsLayerElement is a private class in the Compose UI library so we
will have to check elements' qualified class name to decide if a
modifier element is an instance of it, which is fragile and can be
broken by any reasons. We should remove this once shadows are able to be
faded.

Test: screenshot tests
Bug: 470517507
Change-Id: Ifef49528a7a70103bb7a9f3682e7b55d047f2f2c