Fix incorrect usage of rememberUpdatedState
rememberViewModelStoreNavEntryDecorator is incorrectly using
rememberUpdatedState. The current setup is ignoring the result of
rememberUpdatedState and will never reconstruct the
ViewModelStoreNavEntryDecorator in response to the lambda argument
changing.

This means that users changing the value of
`removeViewModelStoreOnPop` will never see the new lambda argument
used as the first one is captured indefinitely. This CL fixes this by
invoking the latest value from the rememberUpdatedState instance.

Test: testChangeRemoveViewModelStoreOnPop
Change-Id: Ie9576520632150dcd34c1bbf45b6b175d09a2f98