Handle ImageWriter creation failures gracefully
When we fail to create an ImageWriter, a warning log is printed out, but
when an actual reprocessing request is issued later, we fail.

Instead of a warning, print out an error-level log. Also, when
reprocessing request is issued, just abort that request as with any
request that fails to be submitted.

Bug: 470710233
Test: CameraPipe unit tests
Change-Id: Idd9b81999fbf6cb3ba8409c49d9169c78acf9678