Differentiate capture session creation results
As of right now, when capture session creation fails during the process
of CameraDevice.createCaptureSession() calls, we don't differentiate its
return value between a failed call and a successful call with empty
deferred list. The current code works because we make sure to invoke
CaptureSessionState.shutdown() in the synchronously in the same thread.

This CL introduces a dedicated `Result` type, with `Success` and
`Failed` as possible instances. We also correct a few places where
CaptureSessionState.onSessionFinalized() should be called instead of
just shutdown().

Bug: 461602496
Test: Sample app and unit tests
Change-Id: I3ffc20c253f89a7ff6f221414254a29a72e8d1ac