Add Material3TileService that creates correct MaterialScope and ProtoLayoutScope automatically
Add Material3TileService that creates correct MaterialScope and ProtoLayoutScope automatically

With better resources handling and additional M3 and ProtoLayout APIs, ProtoLayoutScope is mandatory. To help developers with having the correct setup when building their Tiles, and not having to worry about creating MaterialScope (as initial long term plans for M3 library), we are adding Material3TileService with only 1 mandatory suspend fun to override, that should produce the layout with resources inlined in it.

This will also restrict usage of other methods that aren't needed/derecated (onTileResourcesRequest, interaction events)

Fixes: 470048768
Test: Added
Relnote: "New, `Material3TileService` has been added to Tiles library. This provides better experience in creating TileService, as it automatically creates `MaterialScope` needed for Material3 components and layout, with `ProtoLayoutScope` for better resource handling and performance improvements. It is more Kotlin friendly, with 1 suspend function to provide both tile layout and resources. Additionally, compile SDK of tiles:tiles library is increased to 35."


Change-Id: I1ff2955c528db8e5133393f52b997e9353272a0f