Add support for generic shapes in stack item decorations.
Generic shapes are inspected to find the widest cross section, which is
used as the bottom of the mask. If there are multiple equal-width cross
sections, then the higher one (on Y axis) is selected.

The logic that selects a decoration in case of equal-width decorations
is updated to select a decoration with a higher widest cross section.

Bug: 467651864
Test: VerticalStackTest, VerticalStackScreenshotTest
Change-Id: I50a2ddb508442fc2aab3ef62d6a8231ea10df82d