Remove a LifecycleCameraRepository warning log
If attempting to unbind use cases from a LifecycleCamera that isn't
managed by the repository, it should be no-op. Currently, a log will be
printed if the camera can't be found in the repository.

In CameraController, it frequently unbind the use cases when
initailizing or reconfiguring, causing a stacking of the log. The
reason is every time the lifecycle owner is updated, it's considered to
be a new LifecycleCamera and the previous LifecycleCamera will be
removed from the repository when the lifecycle is destroyed.

The log is unnecessary because it's expected for the LifecycleCamera to
be removed by the lifecycle event. This CL removes it to aoivd flooding
of the warning log when using CameraController.

Bug: b/466907959
Test: ./gradlew camera:camera-lifecycle:build
Change-Id: I125b4bee853ee2f8ddae7f4c03f68a342bd5cb37