Removes kotlin.metadata from Compose lint checks
This can be replaced with analysis APIs, now that they are stable and are available in most current lint versions. Migrating also fixes an unknown issue in the previous implementation for checking composability of lambdas - some APIs (such as rule.setContent) were incorrectly being marked as _not_ having a composable lambda, because in PSI the lambda parameter name was `p` when in source / kotlin metadata it was `content`, so we didn't match them. Now that we don't need to match parameters in this way, we should have greater coverage, and report more valid errors.

Fixes: b/402490172
Test: lint
Relnote: n/a
Change-Id: Id99fa6cec885b880df20fa5ab66e73c707a2894c